server/requirements.txt
fastapi==0.115.4
pydantic==2.9.2
python-dotenv==1.0.1
redis==5.2.0
Requests==2.32.3
starlette==0.41.2
supabase==2.9.0
uvicorn==0.32.0
